(refactor)refactored server for better code clean up
