+++ server/roms/romsController.js
@@ -1,0 +1,11 @@
+var Roms = require('./romsModel.js');
+var app = require('../server.js');
+
+module.exports = {
+  uniqueGame: function(req, res){
+    var id = req.id;
+    Roms.find({id: id}, function(err, results) {
+      res.send(results);
+    });
+  }
+};
